re #1747 moved ConfigurationImpl from server-base to core to use in ContextClassLoaderFactory
