+++ server/manager/src/main/java/org/apache/accumulo/master/Master.java
@@ -44,7 +44,6 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
-import java.util.function.Supplier;
 
 import org.apache.accumulo.core.Constants;
 import org.apache.accumulo.core.classloader.ContextClassLoaders;
@@ -379,14 +378,7 @@ public class Master extends AbstractServer
     this.tabletBalancer.init(context);
 
     try {
-      ContextClassLoaders.initialize(new Supplier<Map<String,String>>() {
-        @Override
-        public Map<String,String> get() {
-          Map<String,String> props = new HashMap<>();
-          aconf.getProperties(props, key -> !key.isBlank());
-          return props;
-        }
-      });
+      ContextClassLoaders.initialize(aconf);
     } catch (Exception e1) {
       log.error("Error configuring ContextClassLoaderFactory", e1);
       throw new RuntimeException("Error configuring ContextClassLoaderFactory", e1);
