+++ server/tserver/src/main/java/org/apache/accumulo/tserver/TabletServer.java
@@ -51,7 +51,6 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.locks.ReentrantLock;
-import java.util.function.Supplier;
 
 import org.apache.accumulo.core.Constants;
 import org.apache.accumulo.core.classloader.ContextClassLoaders;
@@ -1000,14 +999,7 @@ public class TabletServer extends AbstractServer {
 
     final AccumuloConfiguration aconf = getConfiguration();
     try {
-      ContextClassLoaders.initialize(new Supplier<Map<String,String>>() {
-        @Override
-        public Map<String,String> get() {
-          Map<String,String> props = new HashMap<>();
-          aconf.getProperties(props, key -> !key.isBlank());
-          return props;
-        }
-      });
+      ContextClassLoaders.initialize(aconf);
     } catch (Exception e1) {
       log.error("Error configuring ContextClassLoaderFactory", e1);
       throw new RuntimeException("Error configuring ContextClassLoaderFactory", e1);
