+++ shell/src/main/java/org/apache/accumulo/shell/Shell.java
@@ -44,7 +44,6 @@ import java.util.Properties;
 import java.util.Set;
 import java.util.TreeMap;
 import java.util.concurrent.TimeUnit;
-import java.util.function.Supplier;
 
 import org.apache.accumulo.core.Constants;
 import org.apache.accumulo.core.classloader.ContextClassLoaders;
@@ -62,6 +61,7 @@ import org.apache.accumulo.core.clientImpl.ClientContext;
 import org.apache.accumulo.core.clientImpl.ClientInfo;
 import org.apache.accumulo.core.clientImpl.Tables;
 import org.apache.accumulo.core.conf.ClientProperty;
+import org.apache.accumulo.core.conf.ConfigurationCopy;
 import org.apache.accumulo.core.conf.Property;
 import org.apache.accumulo.core.data.Key;
 import org.apache.accumulo.core.data.Value;
@@ -466,16 +466,8 @@ public class Shell extends ShellOptions implements KeywordExecutable {
 
     if (tableContext != null && !tableContext.equals("")) {
       try {
-        ContextClassLoaders.initialize(new Supplier<Map<String,String>>() {
-          @Override
-          public Map<String,String> get() {
-            try {
-              return shellState.getAccumuloClient().instanceOperations().getSystemConfiguration();
-            } catch (AccumuloException | AccumuloSecurityException e) {
-              throw new RuntimeException("Error getting configuration", e);
-            }
-          }
-        });
+        ContextClassLoaders.initialize(new ConfigurationCopy(
+            shellState.getAccumuloClient().instanceOperations().getSystemConfiguration()));
       } catch (Exception e1) {
         log.error("Error configuring ContextClassLoaderFactory", e1);
         throw new RuntimeException("Error configuring ContextClassLoaderFactory", e1);
